+++ packages/js/src/index.ts
@@ -9,6 +9,9 @@ export { validateNationalID, formatNationalID } from './validators/national-id.j
 export { validateUMID, formatUMID } from './validators/umid.js';
 export { validatePassport, formatPassport } from './validators/passport.js';
 export { validatePRC, formatPRC } from './validators/prc.js';
+export { validateDriversLicense, formatDriversLicense } from './validators/drivers-license.js';
+export { validatePlate, parsePlate } from './validators/plate.js';
+export type { PlateType, PlateParse } from './validators/plate.js';
 
 export { parseMobile, parseLandline, toE164, toNational } from './phone.js';
 export type { MobileParse, LandlineParse, Network } from './phone.js';